PP-5166 Refactor Queue exception handling
