+++ src/main/java/uk/gov/pay/connector/queue/QueueException.java
@@ -1,0 +1,9 @@
+package uk.gov.pay.connector.queue;
+
+
+public class QueueException extends Exception {
+
+    public QueueException(String message) {
+        super(message);
+    }
+}
