--- src/main/java/uk/gov/pay/connector/queue/SqsQueueOperationException.java
@@ -1,9 +1,0 @@
-package uk.gov.pay.connector.queue;
-
-
-public class SqsQueueOperationException extends RuntimeException {
-
-    public SqsQueueOperationException(String message) {
-        super(message);
-    }
-}
